+++ src/main/java/dinkplugin/notifiers/CollectionNotifier.java
@@ -44,6 +44,7 @@ public class CollectionNotifier extends BaseNotifier {
         createMessage(DinkPluginConfig::collectionSendImage, NotificationBody.builder()
             .content(notifyMessage)
             .extra(new CollectionNotificationData(itemName))
+            .screenshotFile("collectionImage.png")
             .type(NotificationType.COLLECTION)
             .build());
     }