+++ src/main/java/dinkplugin/notifiers/PetNotifier.java
@@ -34,6 +34,7 @@ public class PetNotifier extends BaseNotifier {
 
         createMessage(DinkPluginConfig::petSendImage, NotificationBody.builder()
             .content(notifyMessage)
+            .screenshotFile("petImage.png")
             .type(NotificationType.PET)
             .build());
     }