+++ src/main/java/dinkplugin/notifiers/SlayerNotifier.java
@@ -88,6 +88,7 @@ public class SlayerNotifier extends BaseNotifier {
             createMessage(DinkPluginConfig::slayerSendImage, NotificationBody.builder()
                 .content(notifyMessage)
                 .extra(new SlayerNotificationData(slayerTask, slayerCompleted, slayerPoints))
+                .screenshotFile("slayerImage.png")
                 .type(NotificationType.SLAYER)
                 .build());
         }